Move SubmissionProgressView into its own file ahead of work on it.
Move SubmissionProgressView tests into own file.
