+++ backend/audit/cross_validation/__init__.py
@@ -64,7 +64,7 @@ from .check_ref_number_in_findings_text import check_ref_number_in_findings_text
 from .additional_ueis import additional_ueis
 from .auditee_ueis_match import auditee_ueis_match
 from .check_finding_reference_uniqueness import check_finding_reference_uniqueness
-from .sac_validation_shape import sac_validation_shape  # noqa: F401
+from .sac_validation_shape import sac_validation_shape, snake_to_camel  # noqa: F401
 from .submission_progress_check import submission_progress_check
 from .tribal_data_sharing_consent import tribal_data_sharing_consent
